+++ ninedot-api/src/main/java/org/sopt36/ninedotserver/auth/v1/AuthController.java
@@ -1,20 +1,20 @@
-package org.sopt36.ninedotserver.auth.controller;
+package org.sopt36.ninedotserver.auth.v1;
 
-import static org.sopt36.ninedotserver.auth.controller.message.AuthMessage.ACCESS_TOKEN_REFRESH_SUCCESS;
-import static org.sopt36.ninedotserver.auth.controller.message.AuthMessage.LOGIN_SIGNUP_SUCCESS;
-import static org.sopt36.ninedotserver.auth.controller.message.AuthMessage.REFRESH_TOKEN_DELETED;
-import static org.sopt36.ninedotserver.auth.controller.message.AuthMessage.SIGNUP_SUCCESS;
+import static org.sopt36.ninedotserver.auth.v1.message.AuthMessage.ACCESS_TOKEN_REFRESH_SUCCESS;
+import static org.sopt36.ninedotserver.auth.v1.message.AuthMessage.LOGIN_SIGNUP_SUCCESS;
+import static org.sopt36.ninedotserver.auth.v1.message.AuthMessage.REFRESH_TOKEN_DELETED;
+import static org.sopt36.ninedotserver.auth.v1.message.AuthMessage.SIGNUP_SUCCESS;
 
 import jakarta.servlet.http.HttpServletResponse;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
+import org.sopt36.ninedotserver.auth.usecase.AuthService;
 import org.sopt36.ninedotserver.auth.dto.request.GoogleAuthCodeRequest;
+import org.sopt36.ninedotserver.auth.dto.request.SignupRequest;
 import org.sopt36.ninedotserver.auth.dto.response.LoginOrSignupResponse;
 import org.sopt36.ninedotserver.auth.dto.response.NewAccessTokenResponse;
-import org.sopt36.ninedotserver.auth.service.AuthService;
-import org.sopt36.ninedotserver.global.dto.response.ApiResponse;
-import org.sopt36.ninedotserver.auth.dto.request.SignupRequest;
 import org.sopt36.ninedotserver.auth.dto.response.SignupResponse;
+import org.sopt36.ninedotserver.dto.response.ApiResponse;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.CookieValue;
 import org.springframework.web.bind.annotation.PostMapping;
